--- MKSWifiPlugin/plugin.json
@@ -1,16 +1,0 @@
-{
-    "name": "MKS Plugin Connection",
-    "author": "makerbase",
-    "version": "1.1.3",
-    "minimum_cura_version": "4.5",
-    "maximum_cura_version": "4.8",
-    "description": "Сonnect to TFT35 based printers in Cura",
-    "api": "7.1.0",
-    "supported_sdk_versions": [
-        "7.1.0",
-        "7.2.0",
-        "7.3.0",
-        "7.4.0"
-    ],
-    "i18n-catalog": "cura"
-}
